% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_inputs.R
\name{is_valid_one_design}
\alias{is_valid_one_design}
\title{Is a valid design?}
\usage{
is_valid_one_design(design)
}
\arguments{
\item{design}{The design matrix/array}
}
\description{
Is a valid design?
}
\keyword{internal}
