% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_inputs.R
\name{is_valid_one_nuisance}
\alias{is_valid_one_nuisance}
\title{Is a valid nuisance?}
\usage{
is_valid_one_nuisance(nuisance)
}
\arguments{
\item{nuisance}{The nuisance matrix}
}
\description{
Is a valid nuisance?
}
\keyword{internal}
