% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiGLM_fun.R
\name{multiGLM_fun}
\alias{multiGLM_fun}
\title{multiGLM0}
\usage{
multiGLM_fun(
  BOLD,
  design,
  nuisance = NULL,
  design_canonical = NULL,
  verbose = 1,
  meanTol = 1e-06,
  varTol = 1e-06
)
}
\arguments{
\item{BOLD, design, nuisance}{Session-length list of numeric matrices/arrays,
each with volumes along the first dimension.}

\item{design_canonical}{TO DO}

\item{verbose}{\code{1} (default) to print occasional updates during model
computation; \code{2} for occasional updates as well as running INLA in
verbose mode (if \code{Bayes}), or \code{0} for no printed updates.}

\item{meanTol, varTol}{Tolerance for mean, variance and SNR of each data location.
Locations which do not meet these thresholds are masked out of the analysis.
Default: \code{1e-6} for mean and variance, \code{50} for SNR.}
}
\value{
A list with elements
\describe{
\item{bestmodel}{...}
\item{Fstat}{...}
\item{pvalF}{...}
}
}
\description{
Performs classical GLM for task fMRI activation, comparing multiple designs
}
